pyproject.toml
[build-system]
# 40.8.0 is the first version compatible with PEP 517
requires = ["setuptools>=40.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ubermagtable"
version = "0.61.0"
description = "Python package for manipulating tabular data."
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}

authors = [
    {name = "Marijan Beg"},
    {name = "Martin Lang"},
    {name = "Ryan A. Pepper"},
    {name = "Thomas Kluyver"},
    {name = "Samuel Holt"},
    {name = "Hans Fangohr"}
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Education",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Visualization"
]

dependencies = [
    "ubermagutil~=0.61.0",
    "pandas>=1.1",
    "matplotlib>=3.3",
    "jupyterlab~=3.0",
    "ipywidgets>=7.5"
]

[project.optional-dependencies]
dev = [
    "build",
    "invoke",
    "nbval",
    "pre-commit",
    "pytest-cov",
    "twine"
]

[project.urls]
homepage = "https://ubermag.github.io"
documentation = "https://ubermag.github.io/documentation/ubermagtable"
repository = "https://github.com/ubermag/ubermagtable"



# to define additional information we use a custom table
# we need to own the entry for tool.NAME on PyPI
# see: https://www.python.org/dev/peps/pep-0518/#tool-table
[tool.ubermag]
doi = "10.5281/zenodo.3539491"
copyright_holder = "Marijan Beg and Hans Fangohr"
contributors = [
    "Vanessa Nehruji",
    "Sergii Mamedov",
    "Stefan Tibus"
]
about = """
- reading scalar data files (OOMMF `.odt` and mumax3 `.txt`),

- merging multiple tables, and

- visualisation of scalar table data.
"""

[tool.coverage.run]
omit = ["ubermagtable/tests/*"]

[tool.isort]
profile = "black"
skip_gitignore = true  # ignores files listed in .gitignore

[tool.black]
experimental-string-processing = true
